Hyperparameter Helper updates.
Renamed variable "categoricalHelper" to "modelHelper" to be more consistent with the rest of the software.
Removed display output from "RunHypertuning".  Displaying output should be done separately so it can be controlled for individual models.
Simplified HyperparameterHelper unit tests so that models are always saved rather than individually selecting which test cases are saved.